fix: resizing the landing page
