+++ src/app.js
@@ -1,7 +1,8 @@
 import React from 'react';
-import Home from './views/Home/Home';
+import { Home } from './views';
 
 import './App.css';
+// Use bootstrap for styling
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 const App = () => <Home />;
