+++ src/views/Footer/Footer.css
@@ -1,0 +1,23 @@
+.footer {
+  bottom: 0;
+  width: 100%;
+  min-height: 96px;
+  position: absolute;
+  color: rgb(255, 255, 255);
+  background-color: rgb(34, 34, 34);
+}
+
+#footer_content {
+  max-width: 960px;
+}
+
+.footer a,
+.footer a:hover {
+  color: rgb(255, 255, 255);
+  text-decoration: none;
+}
+
+.footer a:hover,
+.footer a:active {
+  text-decoration: underline;
+}
