+++ src/views/Home/Home.css
@@ -1,9 +1,17 @@
+.content {
+  width: 100%;
+  position: absolute;
+  top: 0;
+  bottom: 96px;
+}
+
 #homecontent {
   min-width: 960px;
   padding-top: 96px;
   display: flex;
-  flex-direction: column;
   align-items: center;
+  flex-direction: column;
+  justify-content: space-between;
 }
 
 #navcontent {
@@ -38,11 +46,13 @@
   justify-content: center;
   align-items: center;
   display: flex;
+  padding: 0 0.5rem;
 }
 
 #about {
-  min-height: 96px;
   width: 640px;
+  margin-top: 48px;
+  min-height: 96px;
 }
 
 .trend {
