+++ src/views/index.js
@@ -1,0 +1,4 @@
+import Home from './Home';
+import Footer from './Footer';
+
+export { Home, Footer };
